[translator] move the entry points in index.ts
rename `toolchain.ts` into `index.ts`. The purpose of this module is
precisely to expose translation entry points so that makes sense to just
put everything in index.ts
