+++ src/main.ts
@@ -1,5 +1,5 @@
 export { mongoToPipe } from './lib/pipeline';
-export { getTranslator } from './lib/translators/toolchain';
+export { getTranslator } from './lib/translators';
 
 import Pipeline from './components/Pipeline.vue';
 export { Pipeline };
